beeline_bot/requirements.txt
python-telegram-bot==20.8
pandas
requests
openpyxl